Lots of Kotlin enabled improvements

+++ service-rest/src/main/kotlin/com/bnorm/barkeep/rest/RestBarService.kt
@@ -2,11 +2,9 @@
 package com.bnorm.barkeep.rest
 
 import com.bnorm.barkeep.db.DbBarService
-import com.bnorm.barkeep.db.DbUserService
 import com.bnorm.barkeep.model.Bar
 import com.bnorm.barkeep.model.BarSpec
 import com.bnorm.barkeep.model.Ingredient
-import com.bnorm.barkeep.model.Recipe
 import com.bnorm.barkeep.model.User
 import com.bnorm.barkeep.service.BarService
 import com.fasterxml.jackson.annotation.JsonView
@@ -19,72 +17,89 @@ import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RestController
 
+
 @RestController
 @RequestMapping("/bars")
-class RestBarService(private val userService: DbUserService,
-                     private val barService: DbBarService) : AbstractRestService(), BarService {
+class RestBarService(private val barService: DbBarService,
+                     private val userService: RestUserService,
+                     private val ingredientService: RestIngredientService) : BarService {
 
-  private fun findByOwner(userId: Long, barId: Long): Bar? {
-    val bar = barService.getBar(barId)
-    if (!isOwnedBy(bar, userId)) {
-      throw NotFound("Unable to find bar with id=$barId")
+  private fun isOwnedBy(bar: BarSpec, userId: Long): Boolean = bar.owner?.id == userId
+
+  private fun assertAccess(bar: Bar) {
+    val user = userService.getUser()
+    val bars = user.bars
+    if (!bars.contains(bar)) {
+      throw NotFound(MSG_NOT_FOUND(TYPE_BAR, bar.id))
     }
-    return bar
   }
 
-  private fun findByUser(user: User?, barId: Long): Bar {
-    val bar = barService.getBar(barId)
-    val bars = user?.bars
-    if (bar == null || bars != null && !bars.contains(bar)) {
-      throw NotFound("Unable to find bar with id=$barId")
+  private fun assertOwnership(bar: Bar) {
+    if (!isOwnedBy(bar, userService.currentId())) {
+      throw Forbidden(MSG_DO_NOT_OWN(TYPE_BAR, bar.id))
     }
-    return bar
   }
 
   @JsonView(Bar::class)
   @GetMapping
   override fun getBars(): Collection<Bar> {
-    val user = userService.getUser(currentUser().id)
-    return user!!.bars
+    val user = userService.getUser()
+    return user.bars
   }
 
   @JsonView(Bar::class)
   @GetMapping("/{barId}")
-  override fun getBar(@PathVariable("barId") id: Long): Bar? {
-    val user = userService.getUser(currentUser().id)
-    return findByUser(user, id)
+  override fun getBar(@PathVariable("barId") id: Long): Bar {
+    val bar = barService.findBar(id) ?: throw NotFound(MSG_NOT_FOUND(TYPE_BAR, id))
+    assertAccess(bar)
+    return bar
   }
 
-  @JsonView(Recipe::class)
+  @JsonView(Ingredient::class)
   @GetMapping("/{barId}/ingredients")
-  fun getBookRecipes(@PathVariable("barId") barId: Long): Set<Ingredient> {
+  fun getBarIngredients(@PathVariable("barId") barId: Long): Set<Ingredient> {
     val bar = getBar(barId)
-    return bar!!.ingredients!!
+    return bar.ingredients
   }
 
   @JsonView(Bar::class)
   @PostMapping
   override fun createBar(@RequestBody bar: BarSpec): Bar {
-    val currentUser = currentUser()
+    val currentUser = userService.getUser()
     if (bar.owner == null) {
       return barService.createBar(BarWithOwner(bar, currentUser))
     } else if (!isOwnedBy(bar, currentUser.id)) {
-      throw BadRequest("Cannot create bar owned by another user")
+      throw BadRequest(MSG_CREATE_WRONG_USER("bar"))
     } else {
       return barService.createBar(bar)
     }
   }
 
+  @PostMapping("/{barId}/ingredients/{ingredientId}")
+  override fun addBarIngredient(@PathVariable("barId") barId: Long, @PathVariable("ingredientId") ingredientId: Long) {
+    val bar = getBar(barId)
+    val ingredient = ingredientService.getIngredient(ingredientId)
+    barService.addBarIngredient(bar.id, ingredient.id)
+  }
+
+  @DeleteMapping("/{barId}/ingredients/{ingredientId}")
+  override fun removeBarIngredient(@PathVariable("barId") barId: Long,
+                                   @PathVariable("ingredientId") ingredientId: Long) {
+    val bar = getBar(barId)
+    val ingredient = ingredientService.getIngredient(ingredientId)
+    barService.removeBarIngredient(bar.id, ingredient.id)
+  }
+
   @JsonView(Bar::class)
   @PutMapping("/{barId}")
   override fun setBar(@PathVariable("barId") id: Long, @RequestBody bar: BarSpec): Bar {
-    findByOwner(currentUser().id, id)
+    assertOwnership(getBar(id))
     return barService.setBar(id, bar)
   }
 
   @DeleteMapping("/{barId}")
   override fun deleteBar(@PathVariable("barId") id: Long) {
-    findByOwner(currentUser().id, id)
+    assertOwnership(getBar(id))
     barService.deleteBar(id)
   }
